Modularized the pageInstructor.tsx file to be more manageable and fixed the bugs with the functioning logic.
